Add container name to Hugepage Downward API file names
To support multiple containers in a pod using hugepages, append the container
name to the Hugepage Downward API file names.

Applications running in containers don't have a clean way to determine the
name of the container they are running in, so also add the container name to
an environment variable so hugepages can be processed better.

Resolves: #45

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>