Add github workflows for image creation
With a push to master, a new image will be built and pushed to
github container registry at URL:
ghcr.io/k8snetworkplumbingwg/network-resources-injector:${tag}
The accompanying tag can wither be sha or latest.

With a creation of a new release and associated tag which starts
with 'v', a new image will be built and pushed to github
container registry at URL:
ghcr.io/k8snetworkplumbingwg/network-resources-injector:${tag}
The accompanying tag will start with 'v' and the version number.

Signed-off-by: Kennelly, Martin <martin.kennelly@intel.com>